created words folder and moving all values / parsing code there
