+++ zss/firmware/cli.ts
@@ -9,7 +9,7 @@ import {
 } from 'zss/device/api'
 import { modemwriteinitstring } from 'zss/device/modem'
 import { createfirmware } from 'zss/firmware'
-import { ispresent, isstring, MAYBE } from 'zss/mapping/types'
+import { ispresent, MAYBE } from 'zss/mapping/types'
 import {
   MEMORY_LABEL,
   memoryclearbook,
@@ -25,8 +25,7 @@ import {
 import { bookclearcodepage, bookreadcodepagebyaddress } from 'zss/memory/book'
 import { codepagereadname, codepagereadtypetostring } from 'zss/memory/codepage'
 import { CODE_PAGE, CODE_PAGE_LABEL, CODE_PAGE_TYPE } from 'zss/memory/types'
-
-import { ARG_TYPE, READ_CONTEXT, readargs } from './wordtypes'
+import { ARG_TYPE, READ_CONTEXT, readargs } from 'zss/words/reader'
 
 const ismac = navigator.userAgent.indexOf('Mac') !== -1
 const metakey = ismac ? 'cmd' : 'ctrl'
@@ -229,10 +228,6 @@ export const CLI_FIRMWARE = createfirmware({
       chip.command('pages')
     }
     return 0
-  })
-  .command('pagecreate', (_, words) => {
-    //
-    //
   })
   .command('pageopen', (_, words) => {
     const [page] = readargs(words, 0, [ARG_TYPE.STRING])
