+++ zss/firmware/loader/jsonloader.ts
@@ -1,5 +1,5 @@
 import { FIRMWARE_COMMAND } from 'zss/firmware'
 
-export const jsonloader: FIRMWARE_COMMAND = (chip) => {
+export const jsonloader: FIRMWARE_COMMAND = () => {
   return 0
 }
