+++ zss/firmware/loader/textloader.ts
@@ -1,5 +1,5 @@
 import { FIRMWARE_COMMAND } from 'zss/firmware'
 
-export const textloader: FIRMWARE_COMMAND = (chip) => {
+export const textloader: FIRMWARE_COMMAND = () => {
   return 0
 }
