+++ zss/gadget/components/panel/common.ts
@@ -1,6 +1,6 @@
 import { createContext } from 'react'
 import { WRITE_TEXT_CONTEXT } from 'zss/gadget/data/textformat'
-import { WORD } from 'zss/memory/types'
+import { WORD } from 'zss/words/types'
 
 export type PanelItemProps = {
   player: string
