+++ zss/terminal/manifest.webmanifest
@@ -1,7 +1,7 @@
 {
   "short_name": "zed.cafe",
   "name": "zed.cafe fullscreen",
-  "description": "used to pinned to home",
+  "description": "pin app for fullscreen",
   "scope": "/",
   "background_color": "#000",
   "theme_color": "#00A",
