+++ zss/words/types.ts
@@ -1,0 +1,3 @@
+export type PT = { x: number; y: number }
+export type WORD = string | number | undefined | WORD[]
+export type WORD_RESULT = 0 | 1
